refactor(core): eliminate XxxProvider enums
Replace 13 XxxProvider enums (with Service/Mock variants and delegation
impls) with Box<dyn XxxApi> trait objects in Provider. Consolidate
13 mock.rs files into single mocks.rs with all mock! definitions.

This removes ~3300 lines of repetitive enum, delegation, and per-module
mock boilerplate. Provider fields are now Box<dyn XxxApi>, builder
accepts impl XxxApi for mock injection, and getters return trait refs.

This also slightly improves the compilation time.
