--- crates/core/src/application_credential/mock.rs
@@ -1,72 +1,0 @@
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-use async_trait::async_trait;
-use mockall::mock;
-
-use openstack_keystone_core_types::application_credential::*;
-
-use crate::application_credential::{ApplicationCredentialApi, ApplicationCredentialProviderError};
-use crate::keystone::ServiceState;
-
-mock! {
-    pub ApplicationCredentialProvider {}
-
-    #[async_trait]
-    impl ApplicationCredentialApi for ApplicationCredentialProvider {
-
-        async fn create_access_rule(
-            &self,
-            state: &ServiceState,
-            rule: AccessRuleCreate,
-        ) -> Result<AccessRule, ApplicationCredentialProviderError>;
-
-        async fn create_application_credential(
-            &self,
-            state: &ServiceState,
-            rec: ApplicationCredentialCreate,
-        ) -> Result<ApplicationCredentialCreateResponse, ApplicationCredentialProviderError>;
-
-        async fn delete_access_rule<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            id: &'a str,
-        ) -> Result<(), ApplicationCredentialProviderError>;
-
-        async fn get_access_rule<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            id: &'a str,
-        ) -> Result<Option<AccessRule>, ApplicationCredentialProviderError>;
-
-        async fn get_application_credential<'a>(
-            &self,
-            state: &ServiceState,
-            id: &'a str,
-        ) -> Result<Option<ApplicationCredential>, ApplicationCredentialProviderError>;
-
-        async fn list_access_rules<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<Vec<AccessRule>, ApplicationCredentialProviderError>;
-
-        async fn list_application_credentials(
-            &self,
-            state: &ServiceState,
-            params: &ApplicationCredentialListParameters,
-        ) -> Result<Vec<ApplicationCredential>, ApplicationCredentialProviderError>;
-    }
-}
