+++ crates/core/src/application_credential/mod.rs
@@ -111,217 +111,17 @@
 //!     }
 //! ]
 //! ```
-use async_trait::async_trait;
-
-use openstack_keystone_config::Config;
-use openstack_keystone_core_types::application_credential::*;
 
 pub mod backend;
 pub mod error;
 pub mod hook;
-#[cfg(any(test, feature = "mock"))]
-mod mock;
 mod provider_api;
 pub mod service;
 
-use crate::keystone::ServiceState;
-use crate::plugin_manager::PluginManagerApi;
-use service::ApplicationCredentialService;
-
 pub use error::ApplicationCredentialProviderError;
 pub use hook::ApplicationCredentialHook;
-#[cfg(any(test, feature = "mock"))]
-pub use mock::MockApplicationCredentialProvider;
 pub use provider_api::ApplicationCredentialApi;
+pub use service::ApplicationCredentialService;
 
-/// Application Credential Provider.
-pub enum ApplicationCredentialProvider {
-    Service(ApplicationCredentialService),
-    #[cfg(any(test, feature = "mock"))]
-    Mock(MockApplicationCredentialProvider),
-}
-
-impl ApplicationCredentialProvider {
-    /// Create a new application credential provider.
-    ///
-    /// # Parameters
-    /// - `config`: The service configuration.
-    /// - `plugin_manager`: The plugin manager to retrieve the backend driver.
-    ///
-    /// # Returns
-    /// - `Result<Self, ApplicationCredentialProviderError>` - The created
-    ///   provider or an error.
-    pub fn new<P: PluginManagerApi>(
-        config: &Config,
-        plugin_manager: &P,
-    ) -> Result<Self, ApplicationCredentialProviderError> {
-        Ok(Self::Service(ApplicationCredentialService::new(
-            config,
-            plugin_manager,
-        )?))
-    }
-}
-
-#[async_trait]
-impl ApplicationCredentialApi for ApplicationCredentialProvider {
-    /// Create a standalone access rule owned by a user.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `rule`: The access rule to create (its `user_id` identifies the owner).
-    ///
-    /// # Returns
-    /// - `Result<AccessRule, ApplicationCredentialProviderError>` - The created
-    ///   access rule or an error.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn create_access_rule(
-        &self,
-        state: &ServiceState,
-        rule: AccessRuleCreate,
-    ) -> Result<AccessRule, ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_access_rule(state, rule).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_access_rule(state, rule).await,
-        }
-    }
-
-    /// Create a new application credential.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `rec`: The application credential creation request.
-    ///
-    /// # Returns
-    /// - `Result<ApplicationCredentialCreateResponse,
-    ///   ApplicationCredentialProviderError>` - The creation response or an
-    ///   error.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn create_application_credential(
-        &self,
-        state: &ServiceState,
-        rec: ApplicationCredentialCreate,
-    ) -> Result<ApplicationCredentialCreateResponse, ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_application_credential(state, rec).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_application_credential(state, rec).await,
-        }
-    }
-
-    /// Delete a user's access rule by its ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `user_id`: The ID of the user owning the access rule.
-    /// - `id`: The ID of the access rule.
-    ///
-    /// # Returns
-    /// - `Result<(), ApplicationCredentialProviderError>` - Unit on success, or
-    ///   an error.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn delete_access_rule<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        id: &'a str,
-    ) -> Result<(), ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_access_rule(state, user_id, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_access_rule(state, user_id, id).await,
-        }
-    }
-
-    /// Get a user's access rule by its ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `user_id`: The ID of the user owning the access rule.
-    /// - `id`: The ID of the access rule.
-    ///
-    /// # Returns
-    /// - `Result<Option<AccessRule>, ApplicationCredentialProviderError>` - The
-    ///   access rule if found, or an error.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn get_access_rule<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        id: &'a str,
-    ) -> Result<Option<AccessRule>, ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_access_rule(state, user_id, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_access_rule(state, user_id, id).await,
-        }
-    }
-
-    /// Get a single application credential by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The ID of the application credential.
-    ///
-    /// # Returns
-    /// - `Result<Option<ApplicationCredential>,
-    ///   ApplicationCredentialProviderError>` - A `Result` containing an
-    ///   `Option` with the application credential if found, or an `Error`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn get_application_credential<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<Option<ApplicationCredential>, ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_application_credential(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_application_credential(state, id).await,
-        }
-    }
-
-    /// List all access rules owned by a user.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `user_id`: The ID of the user owning the access rules.
-    ///
-    /// # Returns
-    /// - `Result<Vec<AccessRule>, ApplicationCredentialProviderError>` - A list
-    ///   of access rules or an error.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn list_access_rules<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<Vec<AccessRule>, ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_access_rules(state, user_id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_access_rules(state, user_id).await,
-        }
-    }
-
-    /// List application credentials.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `params`: Parameters for filtering the list of credentials.
-    ///
-    /// # Returns
-    /// - `Result<Vec<ApplicationCredential>,
-    ///   ApplicationCredentialProviderError>` - A list of application
-    ///   credentials or an error.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn list_application_credentials(
-        &self,
-        state: &ServiceState,
-        params: &ApplicationCredentialListParameters,
-    ) -> Result<Vec<ApplicationCredential>, ApplicationCredentialProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_application_credentials(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_application_credentials(state, params).await,
-        }
-    }
-}
+#[cfg(any(test, feature = "mock"))]
+pub use crate::mocks::MockApplicationCredentialProvider;
