+++ crates/core/src/application_credential/provider_api.rs
@@ -25,7 +25,8 @@ pub trait ApplicationCredentialApi: Send + Sync {
     ///
     /// # Parameters
     /// - `state`: The current service state.
-    /// - `rule`: The access rule to create (its `user_id` identifies the owner).
+    /// - `rule`: The access rule to create (its `user_id` identifies the
+    ///   owner).
     ///
     /// # Returns
     /// - `Result<AccessRule, ApplicationCredentialProviderError>` - The created
