+++ crates/core/src/catalog/mod.rs
@@ -30,405 +30,17 @@
 //! An OpenStack service, such as Compute (nova), Object Storage (swift), or
 //! Image service (glance), that provides one or more endpoints through which
 //! users can access resources and perform operations.
-use async_trait::async_trait;
-
-use openstack_keystone_config::Config;
-use openstack_keystone_core_types::catalog::*;
 
 pub mod backend;
 pub mod error;
 pub mod hook;
-#[cfg(any(test, feature = "mock"))]
-mod mock;
 mod provider_api;
 pub mod service;
 
-use crate::keystone::ServiceState;
-use crate::plugin_manager::PluginManagerApi;
-use service::CatalogService;
-
-pub use crate::catalog::error::CatalogProviderError;
+pub use error::CatalogProviderError;
 pub use hook::CatalogHook;
-#[cfg(any(test, feature = "mock"))]
-pub use mock::MockCatalogProvider;
 pub use provider_api::CatalogApi;
+pub use service::CatalogService;
 
-pub enum CatalogProvider {
-    Service(CatalogService),
-    #[cfg(any(test, feature = "mock"))]
-    Mock(MockCatalogProvider),
-}
-
-impl CatalogProvider {
-    /// Creates a new `CatalogProvider`.
-    ///
-    /// # Parameters
-    /// - `config`: The configuration for the catalog provider.
-    /// - `plugin_manager`: The plugin manager used to load the catalog backend.
-    ///
-    /// # Returns
-    /// A `Result` containing the `CatalogProvider` instance or a
-    /// `CatalogProviderError`.
-    pub fn new<P: PluginManagerApi>(
-        config: &Config,
-        plugin_manager: &P,
-    ) -> Result<Self, CatalogProviderError> {
-        Ok(Self::Service(CatalogService::new(config, plugin_manager)?))
-    }
-}
-
-#[async_trait]
-impl CatalogApi for CatalogProvider {
-    /// Create a new endpoint.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `endpoint`: The endpoint creation parameters.
-    ///
-    /// # Returns
-    /// A `Result` containing the created `Endpoint`, or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn create_endpoint(
-        &self,
-        state: &ServiceState,
-        endpoint: EndpointCreate,
-    ) -> Result<Endpoint, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_endpoint(state, endpoint).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_endpoint(state, endpoint).await,
-        }
-    }
-
-    /// Create a new region.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `region`: The region creation parameters.
-    ///
-    /// # Returns
-    /// A `Result` containing the created `Region`, or a `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn create_region(
-        &self,
-        state: &ServiceState,
-        region: RegionCreate,
-    ) -> Result<Region, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_region(state, region).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_region(state, region).await,
-        }
-    }
-
-    /// Create a new service.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `service`: The service creation parameters.
-    ///
-    /// # Returns
-    /// A `Result` containing the created `Service`, or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn create_service(
-        &self,
-        state: &ServiceState,
-        service: ServiceCreate,
-    ) -> Result<Service, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_service(state, service).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_service(state, service).await,
-        }
-    }
-
-    /// Delete an endpoint by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the endpoint.
-    ///
-    /// # Returns
-    /// A `Result` indicating success or a `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn delete_endpoint<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<(), CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_endpoint(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_endpoint(state, id).await,
-        }
-    }
-
-    /// Delete a region by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the region.
-    ///
-    /// # Returns
-    /// A `Result` indicating success or a `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn delete_region<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<(), CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_region(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_region(state, id).await,
-        }
-    }
-
-    /// Delete a service by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the service.
-    ///
-    /// # Returns
-    /// A `Result` indicating success or a `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn delete_service<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<(), CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_service(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_service(state, id).await,
-        }
-    }
-
-    /// Get catalog.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `enabled`: Whether to return only enabled services.
-    ///
-    /// # Returns
-    /// A `Result` containing a vector of tuples of `Service` and its associated
-    /// `Endpoint`s, or a `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn get_catalog(
-        &self,
-        state: &ServiceState,
-        enabled: bool,
-    ) -> Result<Vec<(Service, Vec<Endpoint>)>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_catalog(state, enabled).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_catalog(state, enabled).await,
-        }
-    }
-
-    /// Get single endpoint by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the endpoint.
-    ///
-    /// # Returns
-    /// A `Result` containing an `Option` with the endpoint if found, or an
-    /// `Error`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn get_endpoint<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<Option<Endpoint>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_endpoint(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_endpoint(state, id).await,
-        }
-    }
-
-    /// Get single region by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the region.
-    ///
-    /// # Returns
-    /// A `Result` containing an `Option` with the region if found, or an
-    /// `Error`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn get_region<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<Option<Region>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_region(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_region(state, id).await,
-        }
-    }
-
-    /// Get single service by ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the service.
-    ///
-    /// # Returns
-    /// A `Result` containing an `Option` with the service if found, or an
-    /// `Error`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn get_service<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<Option<Service>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_service(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_service(state, id).await,
-        }
-    }
-
-    /// List Endpoints.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `params`: Parameters for filtering the endpoint list.
-    ///
-    /// # Returns
-    /// A `Result` containing a vector of `Endpoint` objects or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn list_endpoints(
-        &self,
-        state: &ServiceState,
-        params: &EndpointListParameters,
-    ) -> Result<Vec<Endpoint>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_endpoints(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_endpoints(state, params).await,
-        }
-    }
-
-    /// List regions.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `params`: Parameters for filtering the region list.
-    ///
-    /// # Returns
-    /// A `Result` containing a vector of `Region` objects or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn list_regions(
-        &self,
-        state: &ServiceState,
-        params: &RegionListParameters,
-    ) -> Result<Vec<Region>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_regions(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_regions(state, params).await,
-        }
-    }
-
-    /// List services.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `params`: Parameters for filtering the service list.
-    ///
-    /// # Returns
-    /// A `Result` containing a vector of `Service` objects or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn list_services(
-        &self,
-        state: &ServiceState,
-        params: &ServiceListParameters,
-    ) -> Result<Vec<Service>, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_services(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_services(state, params).await,
-        }
-    }
-
-    /// Update an existing endpoint.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the endpoint.
-    /// - `endpoint`: The fields to change.
-    ///
-    /// # Returns
-    /// A `Result` containing the updated `Endpoint`, or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn update_endpoint<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-        endpoint: EndpointUpdate,
-    ) -> Result<Endpoint, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.update_endpoint(state, id, endpoint).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.update_endpoint(state, id, endpoint).await,
-        }
-    }
-
-    /// Update an existing region.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the region.
-    /// - `region`: The fields to change.
-    ///
-    /// # Returns
-    /// A `Result` containing the updated `Region`, or a `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn update_region<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-        region: RegionUpdate,
-    ) -> Result<Region, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.update_region(state, id, region).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.update_region(state, id, region).await,
-        }
-    }
-
-    /// Update an existing service.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The unique identifier of the service.
-    /// - `service`: The fields to change.
-    ///
-    /// # Returns
-    /// A `Result` containing the updated `Service`, or a
-    /// `CatalogProviderError`.
-    #[tracing::instrument(level = "info", skip(self, state))]
-    async fn update_service<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-        service: ServiceUpdate,
-    ) -> Result<Service, CatalogProviderError> {
-        match self {
-            Self::Service(provider) => provider.update_service(state, id, service).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.update_service(state, id, service).await,
-        }
-    }
-}
+#[cfg(any(test, feature = "mock"))]
+pub use crate::mocks::MockCatalogProvider;
