+++ crates/core/src/resource/mod.rs
@@ -30,274 +30,17 @@
 //! A container that groups or isolates resources or identity objects. Depending
 //! on the service operator, a project might map to a customer, account,
 //! organization, or tenant.
-use async_trait::async_trait;
-
-use openstack_keystone_config::Config;
-use openstack_keystone_core_types::resource::*;
 
 pub mod backend;
 pub mod error;
 pub mod hook;
-#[cfg(any(test, feature = "mock"))]
-mod mock;
 mod provider_api;
 pub mod service;
 
-use crate::keystone::ServiceState;
-use crate::plugin_manager::PluginManagerApi;
-use crate::resource::service::ResourceService;
-
-pub use crate::resource::error::ResourceProviderError;
+pub use error::ResourceProviderError;
 pub use hook::ResourceHook;
-#[cfg(any(test, feature = "mock"))]
-pub use mock::MockResourceProvider;
 pub use provider_api::ResourceApi;
+pub use service::ResourceService;
 
-pub enum ResourceProvider {
-    Service(ResourceService),
-    #[cfg(any(test, feature = "mock"))]
-    Mock(MockResourceProvider),
-}
-
-impl ResourceProvider {
-    /// Create a new `ResourceProvider`.
-    ///
-    /// * `config` - The service configuration.
-    /// * `plugin_manager` - The plugin manager used to resolve the backend
-    ///   driver.
-    ///
-    /// A `Result` containing the initialized `ResourceProvider`, or an `Error`.
-    pub fn new<P: PluginManagerApi>(
-        config: &Config,
-        plugin_manager: &P,
-    ) -> Result<Self, ResourceProviderError> {
-        Ok(Self::Service(ResourceService::new(config, plugin_manager)?))
-    }
-}
-
-#[async_trait]
-impl ResourceApi for ResourceProvider {
-    /// Check whether the domain is enabled.
-    ///
-    /// * `state` - The current service state.
-    /// * `domain_id` - The ID of the domain.
-    ///
-    /// A `Result` containing a `bool` indicating whether the domain is enabled,
-    /// or an `Error`.
-    async fn get_domain_enabled<'a>(
-        &self,
-        state: &ServiceState,
-        domain_id: &'a str,
-    ) -> Result<bool, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_domain_enabled(state, domain_id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_domain_enabled(state, domain_id).await,
-        }
-    }
-
-    /// Create a new domain.
-    ///
-    /// * `state` - The current service state.
-    /// * `domain` - The domain details to create.
-    ///
-    /// A `Result` containing the created `Domain`, or an `Error`.
-    async fn create_domain(
-        &self,
-        state: &ServiceState,
-        domain: DomainCreate,
-    ) -> Result<Domain, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_domain(state, domain).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_domain(state, domain).await,
-        }
-    }
-
-    /// Create a new project.
-    ///
-    /// * `state` - The current service state.
-    /// * `project` - The project details to create.
-    ///
-    /// A `Result` containing the created `Project`, or an `Error`.
-    async fn create_project(
-        &self,
-        state: &ServiceState,
-        project: ProjectCreate,
-    ) -> Result<Project, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_project(state, project).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_project(state, project).await,
-        }
-    }
-
-    /// Delete a domain by the ID.
-    ///
-    /// * `state` - The current service state.
-    /// * `id` - The ID of the domain to delete.
-    ///
-    /// A `Result` containing `()` if successful, or an `Error`.
-    async fn delete_domain<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<(), ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_domain(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_domain(state, id).await,
-        }
-    }
-
-    /// Delete a project by the ID.
-    ///
-    /// * `state` - The current service state.
-    /// * `id` - The ID of the project to delete.
-    ///
-    /// A `Result` containing `()` if successful, or an `Error`.
-    async fn delete_project<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<(), ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_project(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_project(state, id).await,
-        }
-    }
-
-    /// Get a single domain.
-    ///
-    /// * `state` - The current service state.
-    /// * `domain_id` - The ID of the domain.
-    ///
-    /// A `Result` containing an `Option` with the `Domain` if found, or an
-    /// `Error`.
-    async fn get_domain<'a>(
-        &self,
-        state: &ServiceState,
-        domain_id: &'a str,
-    ) -> Result<Option<Domain>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_domain(state, domain_id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_domain(state, domain_id).await,
-        }
-    }
-
-    /// Get a single project.
-    ///
-    /// * `state` - The current service state.
-    /// * `project_id` - The ID of the project.
-    ///
-    /// A `Result` containing an `Option` with the `Project` if found, or an
-    /// `Error`.
-    async fn get_project<'a>(
-        &self,
-        state: &ServiceState,
-        project_id: &'a str,
-    ) -> Result<Option<Project>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_project(state, project_id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_project(state, project_id).await,
-        }
-    }
-
-    /// Get a single project by name and domain ID.
-    ///
-    /// * `state` - The current service state.
-    /// * `name` - The name of the project.
-    /// * `domain_id` - The ID of the domain.
-    ///
-    /// A `Result` containing an `Option` with the `Project` if found, or an
-    /// `Error`.
-    async fn get_project_by_name<'a>(
-        &self,
-        state: &ServiceState,
-        name: &'a str,
-        domain_id: &'a str,
-    ) -> Result<Option<Project>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_project_by_name(state, name, domain_id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_project_by_name(state, name, domain_id).await,
-        }
-    }
-
-    /// Get project parents.
-    ///
-    /// * `state` - The current service state.
-    /// * `project_id` - The ID of the project.
-    ///
-    /// A `Result` containing an `Option` with the `Vec<Project>` if found, or
-    /// an `Error`.
-    async fn get_project_parents<'a>(
-        &self,
-        state: &ServiceState,
-        project_id: &'a str,
-    ) -> Result<Option<Vec<Project>>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.get_project_parents(state, project_id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.get_project_parents(state, project_id).await,
-        }
-    }
-
-    /// Find a single domain by its name.
-    ///
-    /// * `state` - The current service state.
-    /// * `domain_name` - The name of the domain.
-    ///
-    /// A `Result` containing an `Option` with the `Domain` if found, or an
-    /// `Error`.
-    async fn find_domain_by_name<'a>(
-        &self,
-        state: &ServiceState,
-        domain_name: &'a str,
-    ) -> Result<Option<Domain>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.find_domain_by_name(state, domain_name).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.find_domain_by_name(state, domain_name).await,
-        }
-    }
-
-    /// List domains.
-    ///
-    /// * `state` - The current service state.
-    /// * `params` - The list parameters.
-    ///
-    /// A `Result` containing a `Vec<Domain>`, or an `Error`.
-    async fn list_domains(
-        &self,
-        state: &ServiceState,
-        params: &DomainListParameters,
-    ) -> Result<Vec<Domain>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_domains(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_domains(state, params).await,
-        }
-    }
-
-    /// List projects.
-    ///
-    /// * `state` - The current service state.
-    /// * `params` - The list parameters.
-    ///
-    /// A `Result` containing a `Vec<Project>`, or an `Error`.
-    async fn list_projects(
-        &self,
-        state: &ServiceState,
-        params: &ProjectListParameters,
-    ) -> Result<Vec<Project>, ResourceProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_projects(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_projects(state, params).await,
-        }
-    }
-}
+#[cfg(any(test, feature = "mock"))]
+pub use crate::mocks::MockResourceProvider;
