--- crates/core/src/revoke/mock.rs
@@ -1,48 +1,0 @@
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-//! Token revocation - internal mocking tools.
-use async_trait::async_trait;
-use mockall::mock;
-
-use openstack_keystone_core_types::revoke::*;
-use openstack_keystone_core_types::token::FernetToken;
-
-use crate::auth::ValidatedSecurityContext;
-use crate::keystone::ServiceState;
-use crate::revoke::{RevokeApi, RevokeProviderError};
-
-mock! {
-    pub RevokeProvider {}
-
-    #[async_trait]
-    impl RevokeApi for RevokeProvider {
-        async fn create_revocation_event(
-            &self,
-            state: &ServiceState,
-            event: RevocationEventCreate
-        ) -> Result<RevocationEvent, RevokeProviderError>;
-
-        async fn is_token_revoked(
-            &self,
-            state: &ServiceState,
-            token_security_context: &ValidatedSecurityContext,
-        ) -> Result<bool, RevokeProviderError>;
-
-        async fn revoke_token(
-            &self,
-            state: &ServiceState,
-            token: &FernetToken,
-        ) -> Result<(), RevokeProviderError>;
-    }
-}
