--- crates/core/src/role/mock.rs
@@ -1,96 +1,0 @@
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-use async_trait::async_trait;
-use mockall::mock;
-
-use openstack_keystone_core_types::role::*;
-
-use crate::keystone::ServiceState;
-use crate::role::{RoleApi, RoleProviderError};
-
-mock! {
-    pub RoleProvider {}
-
-    #[async_trait]
-    impl RoleApi for RoleProvider {
-        async fn create_role(
-            &self,
-            state: &ServiceState,
-            params: RoleCreate,
-        ) -> Result<Role, RoleProviderError>;
-
-        async fn create_role_imply_rule<'a>(
-            &self,
-            state: &ServiceState,
-            prior_role_id: &'a str,
-            implied_role_id: &'a str,
-        ) -> Result<RoleImply, RoleProviderError>;
-
-        async fn check_role_imply_rule<'a>(
-            &self,
-            state: &ServiceState,
-            prior_role_id: &'a str,
-            implied_role_id: &'a str,
-        ) -> Result<bool, RoleProviderError>;
-
-        async fn delete_role<'a>(
-            &self,
-            state: &ServiceState,
-            id: &'a str,
-        ) -> Result<(), RoleProviderError>;
-
-        async fn delete_role_imply_rule<'a>(
-            &self,
-            state: &ServiceState,
-            prior_role_id: &'a str,
-            implied_role_id: &'a str,
-        ) -> Result<(), RoleProviderError>;
-
-        async fn expand_implied_roles(
-            &self,
-            state: &ServiceState,
-            roles: &mut Vec<RoleRef>,
-        ) -> Result<(), RoleProviderError>;
-
-        async fn get_role<'a>(
-            &self,
-            state: &ServiceState,
-            id: &'a str,
-        ) -> Result<Option<Role>, RoleProviderError>;
-
-        async fn get_role_imply_rule<'a>(
-            &self,
-            state: &ServiceState,
-            prior_role_id: &'a str,
-            implied_role_id: &'a str,
-        ) -> Result<Option<RoleImply>, RoleProviderError>;
-
-        async fn list_role_imply_rules(
-            &self,
-            state: &ServiceState,
-        ) -> Result<Vec<RoleImply>, RoleProviderError>;
-
-        async fn list_role_imply_rules_by_prior<'a>(
-            &self,
-            state: &ServiceState,
-            prior_role_id: &'a str,
-        ) -> Result<Vec<RoleImply>, RoleProviderError>;
-
-        async fn list_roles(
-            &self,
-            state: &ServiceState,
-            params: &RoleListParameters,
-        ) -> Result<Vec<Role>, RoleProviderError>;
-    }
-}
