+++ crates/core/src/token/mod.rs
@@ -19,270 +19,21 @@
 //! that does not aspire to be a full-fledged identity store and management
 //! solution.
 
-use async_trait::async_trait;
-
-use openstack_keystone_config::Config;
-use openstack_keystone_core_types::auth::{ScopeInfo, SecurityContext};
 pub use openstack_keystone_core_types::token::*;
 
 pub mod backend;
 pub mod error;
 pub mod hook;
-#[cfg(any(test, feature = "mock"))]
-mod mock;
 mod provider_api;
 pub mod service;
 
-use crate::auth::ValidatedSecurityContext;
-use crate::keystone::ServiceState;
-use crate::plugin_manager::PluginManagerApi;
-use crate::token::service::TokenService;
 pub use error::TokenProviderError;
 pub use hook::TokenHook;
 pub use provider_api::TokenApi;
+pub use service::TokenService;
 
 #[cfg(any(test, feature = "mock"))]
-pub use mock::MockTokenProvider;
-
-pub enum TokenProvider {
-    Service(TokenService),
-    #[cfg(any(test, feature = "mock"))]
-    Mock(MockTokenProvider),
-}
-
-impl TokenProvider {
-    /// Creates a new `TokenProvider` instance.
-    ///
-    /// # Parameters
-    /// - `config`: The system configuration.
-    /// - `plugin_manager`: The plugin manager to resolve backends.
-    ///
-    /// # Returns
-    /// - `Result<Self, TokenProviderError>` - The new provider instance or an
-    ///   error.
-    pub fn new<P: PluginManagerApi>(
-        config: &Config,
-        plugin_manager: &P,
-    ) -> Result<Self, TokenProviderError> {
-        Ok(Self::Service(TokenService::new(config, plugin_manager)?))
-    }
-}
-
-#[async_trait]
-impl TokenApi for TokenProvider {
-    /// Authenticate by token.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `credential`: The token credential string.
-    /// - `allow_expired`: Whether to allow expired tokens.
-    /// - `window_seconds`: Expiration buffer in seconds.
-    ///
-    /// # Returns
-    /// - `Result<AuthenticatedInfo, TokenProviderError>` - Authenticated
-    ///   information or an error.
-    #[tracing::instrument(level = "info", skip(self, state, credential), err(Debug))]
-    async fn authorize_by_token<'a>(
-        &self,
-        state: &ServiceState,
-        credential: &'a str,
-        allow_expired: Option<bool>,
-        window_seconds: Option<i64>,
-    ) -> Result<ValidatedSecurityContext, TokenProviderError> {
-        match self {
-            Self::Service(provider) => {
-                provider
-                    .authorize_by_token(state, credential, allow_expired, window_seconds)
-                    .await
-            }
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => {
-                provider
-                    .authorize_by_token(state, credential, allow_expired, window_seconds)
-                    .await
-            }
-        }
-    }
-
-    /// Validate the token and produce a [`ValidatedSecurityContext`].
-    #[tracing::instrument(level = "info", skip(self, state, credential), err(Debug))]
-    async fn validate_to_context<'a>(
-        &self,
-        state: &ServiceState,
-        credential: &'a str,
-        allow_expired: Option<bool>,
-        window_seconds: Option<i64>,
-    ) -> Result<ValidatedSecurityContext, TokenProviderError> {
-        match self {
-            Self::Service(provider) => {
-                provider
-                    .validate_to_context(state, credential, allow_expired, window_seconds)
-                    .await
-            }
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => {
-                provider
-                    .validate_to_context(state, credential, allow_expired, window_seconds)
-                    .await
-            }
-        }
-    }
-
-    /// Issue a token and produce a [`ValidatedSecurityContext`].
-    #[tracing::instrument(level = "debug", skip(self, state), err(Debug))]
-    async fn issue_token_context(
-        &self,
-        state: &ServiceState,
-        ctx: &SecurityContext,
-        scope: &ScopeInfo,
-    ) -> Result<ValidatedSecurityContext, TokenProviderError> {
-        match self {
-            Self::Service(provider) => provider.issue_token_context(state, ctx, scope).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.issue_token_context(state, ctx, scope).await,
-        }
-    }
-
-    /// Encode the token into a `String` representation.
-    ///
-    /// # Parameters
-    /// - `token`: The token to encode.
-    ///
-    /// # Returns
-    /// - `Result<String, TokenProviderError>` - The encoded string or an error.
-    fn encode_token(&self, token: &FernetToken) -> Result<String, TokenProviderError> {
-        match self {
-            Self::Service(provider) => provider.encode_token(token),
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.encode_token(token),
-        }
-    }
-
-    /// Get the token restriction by the ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The restriction ID.
-    /// - `expand_roles`: Whether to expand roles.
-    ///
-    /// # Returns
-    /// - `Result<Option<TokenRestriction>, TokenProviderError>` - A `Result`
-    ///   containing an `Option` with the token restriction if found, or an
-    ///   `Error`.
-    async fn get_token_restriction<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-        expand_roles: bool,
-    ) -> Result<Option<TokenRestriction>, TokenProviderError> {
-        match self {
-            Self::Service(provider) => {
-                provider
-                    .get_token_restriction(state, id, expand_roles)
-                    .await
-            }
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => {
-                provider
-                    .get_token_restriction(state, id, expand_roles)
-                    .await
-            }
-        }
-    }
-
-    /// Create new token restriction.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `restriction`: The restriction data to create.
-    ///
-    /// # Returns
-    /// - `Result<TokenRestriction, TokenProviderError>` - The created token
-    ///   restriction or an error.
-    async fn create_token_restriction<'a>(
-        &self,
-        state: &ServiceState,
-        restriction: TokenRestrictionCreate,
-    ) -> Result<TokenRestriction, TokenProviderError> {
-        match self {
-            Self::Service(provider) => provider.create_token_restriction(state, restriction).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.create_token_restriction(state, restriction).await,
-        }
-    }
-
-    /// List token restrictions.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `params`: Parameters for listing restrictions.
-    ///
-    /// # Returns
-    /// - `Result<Vec<TokenRestriction>, TokenProviderError>` - A list of token
-    ///   restrictions or an error.
-    async fn list_token_restrictions<'a>(
-        &self,
-        state: &ServiceState,
-        params: &TokenRestrictionListParameters,
-    ) -> Result<Vec<TokenRestriction>, TokenProviderError> {
-        match self {
-            Self::Service(provider) => provider.list_token_restrictions(state, params).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.list_token_restrictions(state, params).await,
-        }
-    }
-
-    /// Update existing token restriction.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The restriction ID.
-    /// - `restriction`: The update data.
-    ///
-    /// # Returns
-    /// - `Result<TokenRestriction, TokenProviderError>` - The updated token
-    ///   restriction or an error.
-    async fn update_token_restriction<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-        restriction: TokenRestrictionUpdate,
-    ) -> Result<TokenRestriction, TokenProviderError> {
-        match self {
-            Self::Service(provider) => {
-                provider
-                    .update_token_restriction(state, id, restriction)
-                    .await
-            }
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => {
-                provider
-                    .update_token_restriction(state, id, restriction)
-                    .await
-            }
-        }
-    }
-
-    /// Delete token restriction by the ID.
-    ///
-    /// # Parameters
-    /// - `state`: The current service state.
-    /// - `id`: The restriction ID.
-    ///
-    /// # Returns
-    /// - `Result<(), TokenProviderError>` - Ok on success, or an error.
-    async fn delete_token_restriction<'a>(
-        &self,
-        state: &ServiceState,
-        id: &'a str,
-    ) -> Result<(), TokenProviderError> {
-        match self {
-            Self::Service(provider) => provider.delete_token_restriction(state, id).await,
-            #[cfg(any(test, feature = "mock"))]
-            Self::Mock(provider) => provider.delete_token_restriction(state, id).await,
-        }
-    }
-}
+pub use crate::mocks::MockTokenProvider;
 
 #[cfg(test)]
 mod tests {
@@ -294,7 +45,6 @@ mod tests {
 
     pub(super) fn setup_config() -> Config {
         let keys_dir = tempdir().unwrap();
-        // write fernet key used to generate tokens in python
         let file_path = keys_dir.path().join("0");
         let mut tmp_file = File::create(file_path).unwrap();
         write!(tmp_file, "BFTs1CIVIBLTP4GOrQ26VETrJ7Zwz1O4wbEcCQ966eM=").unwrap();
